core: 64 bit compatibility
Only minor changes are required to make the kernel 64 bit compatible.
Most of the changes are either DEBUG/printf formatting or different types for void pointer casting.

The only other change is the type of the `data` member in priority_queue_node_t, as `data` must be able to store a pointer.
For current architectures, the assumption `sizeof(unsigned int) == sizeof(void *)` holds, but not for 64 bit.
Therefore, the type is changed to `uintptr_t', which has the same size for the current architectures, but can also store a pointer in 64 bits.

+++ core/lib/include/priority_queue.h
@@ -32,7 +32,7 @@ extern "C" {
 typedef struct priority_queue_node {
     struct priority_queue_node *next;   /**< next queue node */
     uint32_t priority;                  /**< queue node priority */
-    unsigned int data;                  /**< queue node data */
+    uintptr_t data;                     /**< queue node data */
 } priority_queue_node_t;
 
 /**
